error Handler

+++ src/anime/anime.controller.js
@@ -14,12 +14,17 @@ exports.deleteManyByUid = async (req, res) => {
   await AnimeModel.deleteMany({ uid: { $in: req.body.uid } });
   return res.sendStatus(204);
 };
+
 exports.deleteOneByUid = async (req, res) => {
+  const anime = await AnimeModel.findOne({ uid: req.body.uid });
+  if (!anime) throw new Error("NOT_FOUND");
   await AnimeModel.deleteOne({ uid: req.body.uid });
   return res.sendStatus(204);
 };
 
 exports.updateOne = async (req, res) => {
+  const anime = await AnimeModel.findOne({ uid: req.body.uid });
+  if (!anime) throw new Error("NOT_FOUND");
   await AnimeModel.updateOne({ uid: req.body.uid }, req.body);
   return res.sendStatus(200);
 };
@@ -33,13 +38,16 @@ exports.updateMany = async (req, res) => {
 };
 
 exports.readOne = async (req, res) => {
-  const anime = await AnimeModel.find({ uid: req.body.uid }).populate({path : 'reviews', options:{sort :{score : -1}}}).lean();
-  
+  const anime = await AnimeModel.findOne({ uid: req.body.uid })
+    .populate({ path: "reviews", options: { sort: { score: -1 } } })
+    .lean();
+  if (!anime) throw new Error("NOT_FOUND");
+
   return res.status(200).json(anime);
 };
 
-
 exports.readMany = async (req, res) => {
   const anime = await AnimeModel.find({ uid: { $in: req.body.uid } });
+  if (!anime) throw new Error("NOT_FOUND");
   return res.status(200).json(anime);
 };
